pyproject.toml
[tool.poetry]
name = "internet_computer-tools"
version = "0.1.0"
description = "InternetComputer Tools provides tools for querying and working with the Internet Computer."
authors = ["Ryan Cammer <ryancammer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.2.0"
PyYAML = "^6.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
